feat: create winners-store and move all stores in store folder
