+++ src/app/store/winners-store.ts
@@ -1,0 +1,3 @@
+import { createStore } from '@ripetchor/ssm';
+
+export const winnersStore = createStore({ currentPage: 1, totalCount: 0 });
